cloudrun/crf-signedurl-adv/requirements.txt
Flask==2.0.2
functions-framework==3.0.0
google-cloud-storage
Werkzeug==2.2.2